refactor(runtime): extract model provider
