+++ .changeset/extract-runtime-model-provider.md
@@ -1,0 +1,5 @@
+---
+"@opengeni/runtime": patch
+---
+
+Extract provider client construction, request policy, and model routing into a package-private runtime module while preserving the existing public entrypoint and exports.
